+++ doc/CHANGELOG.rst
@@ -4,8 +4,13 @@ Changelog
 This changelog *only* contains changes from the *first* pypi release (0.5.4.3) onwards.
 
 Latest Changes:
-
 - **1.0.3_dev0 - unreleased**
+
+  - Ctrl+C behavior is switchable with interrupt flag to startJVM.
+    If True, process will halt on Ctrl-C.  If False, the process
+    will transfer control to Python rather than halting.  If
+    not specified JPype will assume false if Python is  started as an
+    interactive shell.
 
   - Fixed crash when manually creating wrappers for anonymous classes.